+++ MoviesApp/MovieAPIService.swift
@@ -12,11 +12,6 @@ protocol MovieAPIServiceProtocol {
 }
 
 final class MovieAPIService: MovieAPIServiceProtocol {
-    // MARK: - Private  Properties
-
-    private let session = URLSession.shared
-    private let jsonDecoder = JSONDecoder()
-
     // MARK: - Public Methods
 
     func getDecodable<T>(
@@ -29,7 +24,7 @@ final class MovieAPIService: MovieAPIServiceProtocol {
             return
         }
 
-        session.dataTask(with: url) { data, _, error in
+        URLSession.shared.dataTask(with: url) { data, _, error in
             var result: Result<T?, Error>
 
             defer {
@@ -45,36 +40,13 @@ final class MovieAPIService: MovieAPIServiceProtocol {
 
             guard
                 let data = data,
-                let page = try? self.jsonDecoder.decode(T.self, from: data)
+                let page = try? JSONDecoder().decode(T.self, from: data)
             else {
                 result = .success(nil)
                 return
             }
 
             result = .success(page)
-        }.resume()
-    }
-
-    func downloadImage(url: URL, completion: @escaping (Result<Data?, Error>) -> ()) {
-        session.dataTask(with: url) { data, _, error in
-            var result: Result<Data?, Error>
-
-            defer {
-                DispatchQueue.main.async {
-                    completion(result)
-                }
-            }
-
-            if let error = error {
-                result = .failure(error)
-                return
-            }
-
-            guard let data = data else {
-                result = .success(nil)
-                return
-            }
-            result = .success(data)
         }.resume()
     }
 }
