+++ MoviesApp/MoviesList/ViewModel/MovieCellViewModel.swift
@@ -16,7 +16,7 @@ final class MovieCellViewModel: MovieCellViewModelProtocol {
     // MARK: - Private Properties
 
     private let cacheImageService = CacheImageService()
-    private var networkService = MovieAPIService()
+    private var imageService = ImageAPIService()
 
     // MARK: - Public Methods
 
@@ -32,7 +32,7 @@ final class MovieCellViewModel: MovieCellViewModelProtocol {
             return
         }
 
-        networkService.downloadImage(url: posterURL) { [weak self] result in
+        imageService.downloadImage(url: posterURL) { [weak self] result in
             guard let self = self else { return }
             switch result {
             case let .success(data):
